+++ tests/decoding_tests.py
@@ -2,27 +2,22 @@
 import decimal
 import struct
 import time
-try:
-    import unittest2 as unittest
-except ImportError:
-    import unittest
+import unittest
 
-from pamqp import decode
-from pamqp import PYTHON3
+from pamqp import decode, PYTHON3
 
 if PYTHON3:
     long = int
-    unicode = bytes
-    basestring = bytes
+    unicode = None
+    basestring = (bytes, str)
 
 PLATFORM_32BIT = (struct.calcsize('P') * 8) == 32
 PLATFORM_64BIT = (struct.calcsize('P') * 8) == 64
 
 
-def utf8(value):
-    if PYTHON3:
-        return value
-    return value.decode('utf-8')
+def _utf8_key(value):
+    return value.decode('utf-8') if not PYTHON3 else value
+
 
 
 class CodecDecodeTests(unittest.TestCase):
@@ -37,22 +32,17 @@ class CodecDecodeTests(unittest.TestCase):
                        3.14,
                        3294967295,
                        -9223372036854775800]
-    FIELD_TBL = (b'\x00\x00\x00\x97\x08arrayvalA\x00\x00\x00\ts\x00'
-                 b'\x01s\x00\x02s\x00\x03\x07boolvalt\x01\x06decvalD'
-                 b'\x02\x00\x00\x01:\x07dictvalF\x00\x00\x00\r\x04f'
-                 b'\xe2\x9c\x89S\x00\x00\x00\x03\xe2\x9c\x90\x08'
-                 b'floatvalf@H\xf5\xc3\x06intvals\x00\x01\x07'
-                 b'longvalI6e&U\x06strvalS\x00\x00\x00\x08Test \xe2'
-                 b'\x9c\x88\x0ctimestampvalT\x00\x00\x00\x00Ec)\x92')
+    FIELD_TBL = (b'\x00\x00\x00\x9d\x08arrayvalA\x00\x00\x00\ts\x00\x01s\x00\x02s\x00\x03\x07boolvalt\x01\x06decvalD\x02\x00\x00\x01:\x07dictvalF\x00\x00\x00\r\x04f\xe2\x9c\x89S\x00\x00\x00\x03\xe2\x9c\x90\x08floatvalf@H\xf5\xc3\x06intvals\x00\x01\x07longvalI6e&U\x06strvalS\x00\x00\x00\x08Test \xe2\x9c\x88\x0ctimestampvalT\x00\x00\x00\x00Ec)\x92\x04\xf0\x9f\x90\xb0V')
     FIELD_TBL_VALUE = {'intval': 1,
                        'strval': b'Test \xe2\x9c\x88',
                        'boolval': True,
                        'timestampval': time.struct_time((2006, 11, 21, 16, 30,
                                                          10, 1, 325, 0)),
                        'decval': decimal.Decimal('3.14'),
+                       _utf8_key('🐰'): None,
                        'floatval': 3.14,
                        'longval': long(912598613),
-                       'dictval': {utf8('f✉'): b'\xe2\x9c\x90'},
+                       'dictval': {_utf8_key('f✉'): b'\xe2\x9c\x90'},
                        'arrayval': [1, 2, 3]}
 
     def test_decode_by_type_invalid_data_type(self):
@@ -92,6 +82,9 @@ class CodecDecodeTests(unittest.TestCase):
         self.assertEqual(decode.byte_array(b'\x00\x00\x00\x03ABC'),
                          (7, bytearray([65, 66, 67])))
 
+    def test_decode_byte_array_invalid_value(self):
+        self.assertRaises(ValueError, decode.byte_array, None)
+
     def test_decode_decimal_value_bytes_consumed(self):
         value = b'\x05\x00\x04\xcb/'
         self.assertEqual(decode.decimal(value)[0], len(value))
@@ -107,6 +100,14 @@ class CodecDecodeTests(unittest.TestCase):
 
     def test_decode_decimal_invalid_value(self):
         self.assertRaises(ValueError, decode.decimal, False)
+
+    def test_decode_double_value(self):
+        value = b'@\t!\xf9\xf0\x1b\x86n'
+        self.assertEqual(round(decode.double(value)[1], 5),
+                         round(float(3.14159), 5))
+
+    def test_decode_double_invalid_value(self):
+        self.assertRaises(ValueError, decode.double, 123)
 
     def test_decode_embedded_value_null(self):
         self.assertEqual(decode._embedded_value(b'\00')[1], None)
@@ -150,12 +151,12 @@ class CodecDecodeTests(unittest.TestCase):
         self.assertEqual(decode.long_long_int(value)[0], 8)
 
     @unittest.skipIf(PLATFORM_32BIT, 'Skipped on 32-bit platforms')
-    def test_decode_long_long_int_data_type(self):
+    def test_decode_long_long_int_data_type_64bit(self):
         value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertIsInstance(decode.long_long_int(value)[1], int)
 
     @unittest.skipIf(PLATFORM_64BIT, 'Skipped on 64-bit platforms')
-    def test_decode_long_long_int_data_type(self):
+    def test_decode_long_long_int_data_type_32bit(self):
         value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertIsInstance(decode.long_long_int(value)[1], long)
 
@@ -218,6 +219,15 @@ class CodecDecodeTests(unittest.TestCase):
         value = b'\x7f\xff'
         self.assertEqual(decode.short_int(value)[1], 32767)
 
+    def test_decode_short_short_int_data_type(self):
+        self.assertIsInstance(decode.short_short_int(b'\xff')[0], int)
+
+    def test_decode_short_short_int_invalid_value(self):
+        self.assertRaises(ValueError, decode.short_short_int, None)
+
+    def test_decode_short_short_int_value(self):
+        self.assertEqual(decode.short_short_int(b'\xff')[1], 255)
+
     def test_decode_short_str_bytes_consumed(self):
         self.assertEqual(decode.short_str(b'\n0123456789')[0], 11)
 
@@ -279,7 +289,7 @@ class CodecDecodeTests(unittest.TestCase):
 
     def test_decode_field_table_value(self):
         value = decode.field_table(self.FIELD_TBL)[1]
-        for key in value.keys():
+        for key in self.FIELD_TBL_VALUE.keys():
             if isinstance(value[key], float):
                 self.assertAlmostEqual(round(value[key], 3),
                                        round(self.FIELD_TBL_VALUE[key], 3))
@@ -316,6 +326,20 @@ class CodecDecodeTests(unittest.TestCase):
     def test_decode_by_type_boolean_true_data_type(self):
         self.assertIsInstance(decode.by_type(b'\x01', 'boolean')[1], bool)
 
+    def test_decode_by_type_byte_array_bytes_consumed(self):
+        value = b'\x00\x00\x00\t123456789'
+        self.assertEqual(decode.by_type(value, 'byte_array')[0], 13)
+
+    def test_decode_by_type_byte_array_data_type(self):
+        value = b'\x00\x00\x00\t123456789'
+        self.assertIsInstance(decode.by_type(value, 'byte_array')[1],
+                              bytearray)
+
+    def test_decode_by_type_byte_array_value(self):
+        value = b'\x00\x00\x00\t123456789'
+        self.assertEqual(decode.by_type(value, 'byte_array')[1],
+                         bytearray(b'123456789'))
+
     def test_decode_by_type_decimal_bytes_consumed(self):
         value = b'\x05\x00\x04\xcb/'
         self.assertEqual(decode.by_type(value, 'decimal')[0], len(value))
@@ -333,13 +357,26 @@ class CodecDecodeTests(unittest.TestCase):
     def test_decode_by_type_decimal_invalid_value(self):
         self.assertRaises(ValueError, decode.by_type, False, 'decimal')
 
-    def test_decode_by_type_floating_point_bytes_consumed(self):
-        self.assertEqual(decode.by_type(b'@I\x0f\xd0', 'float')[0], 4)
+    def test_decode_by_type_double_data_type(self):
+        value = b'C\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertIsInstance(decode.by_type(value, 'double')[1],
+                              float)
+
+    def test_decode_by_type_double_bytes_consumed(self):
+        value = b'C\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertEqual(decode.by_type(value, 'double')[0], 8)
+
+    def test_decode_by_type_double_value(self):
+        value = b'C\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertEqual(decode.by_type(value, 'double')[1],
+                         1120480238450803.0)
 
     def test_decode_by_type_floating_point_data_type(self):
-        value = b'@I\x0f\xd0'
-        self.assertIsInstance(decode.by_type(value, 'float')[1],
+        self.assertIsInstance(decode.by_type(b'@I\x0f\xd0', 'float')[1],
                               float)
+
+    def test_decode_by_type_float_bytes_consumed(self):
+        self.assertEqual(decode.by_type(b'@I\x0f\xd0', 'float')[0], 4)
 
     def test_decode_by_type_floating_point_invalid_value(self):
         self.assertRaises(ValueError, decode.by_type, False, 'float')
@@ -368,15 +405,15 @@ class CodecDecodeTests(unittest.TestCase):
         value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertEqual(decode.by_type(value, 'longlong')[0], 8)
 
-    @unittest.skipIf(PLATFORM_32BIT, 'Skipped on 32-bit platforms')
-    def test_decode_by_type_long_long_data_type(self):
-        value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
-        self.assertIsInstance(decode.by_type(value, 'longlong')[1], int)
-
     @unittest.skipIf(PLATFORM_64BIT, 'Skipped on 64-bit platforms')
-    def test_decode_by_type_long_long_data_type(self):
+    def test_decode_by_type_long_long_data_type_32bit(self):
         value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertIsInstance(decode.by_type(value, 'longlong')[1], long)
+
+    @unittest.skipIf(PLATFORM_32BIT, 'Skipped on 32-bit platforms')
+    def test_decode_by_type_long_long_data_type_64bit(self):
+        value = b'\x7f\xff\xff\xff\xff\xff\xff\xf8'
+        self.assertIsInstance(decode.by_type(value, 'longlong')[1], int)
 
     def test_decode_by_type_long_long_invalid_value(self):
         self.assertRaises(ValueError, decode.by_type, None, 'longlong')
@@ -450,6 +487,9 @@ class CodecDecodeTests(unittest.TestCase):
                          time.struct_time((2006, 11, 21,
                                            16, 30, 10, 1, 325, 0)))
 
+    def test_decode_by_type_void(self):
+        self.assertIsNone(decode.by_type(b'', 'void'))
+
     def test_decode_by_type_field_array_bytes_consumed(self):
         self.assertEqual(decode.by_type(self.FIELD_ARR, 'array')[0],
                          len(self.FIELD_ARR))
@@ -485,7 +525,7 @@ class CodecDecodeTests(unittest.TestCase):
 
     def test_decode_by_type_field_table_value(self):
         value = decode.by_type(self.FIELD_TBL, 'table')[1]
-        for key in value.keys():
+        for key in self.FIELD_TBL_VALUE.keys():
             if isinstance(value[key], float):
                 self.assertAlmostEqual(round(value[key], 3),
                                        round(self.FIELD_TBL_VALUE[key], 3))
@@ -512,6 +552,20 @@ class CodecDecodeTests(unittest.TestCase):
         self.assertEqual(round(float(decode._embedded_value(value)[1]), 5),
                          round(float(decimal.Decimal('3.14159')), 5))
 
+    def test_decode_embedded_value_double_bytes_consumed(self):
+        value = b'dC\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertEqual(decode._embedded_value(value)[0], len(value))
+
+    def test_decode_embedded_value_double_data_type(self):
+        value = b'dC\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertIsInstance(decode._embedded_value(value)[1],
+                              float)
+
+    def test_decode_embedded_value_double_value(self):
+        value = b'dC\x0f\xd8\x91\x14\xb9\xc3\x98'
+        self.assertEqual(decode._embedded_value(value)[1],
+                         1120480238450803.0)
+
     def test_decode_embedded_value_long_bytes_consumed(self):
         value = b'I\x7f\xff\xff\xff'
         self.assertEqual(decode._embedded_value(value)[0], 5)
@@ -524,17 +578,29 @@ class CodecDecodeTests(unittest.TestCase):
         value = b'I\x7f\xff\xff\xff'
         self.assertEqual(decode._embedded_value(value)[1], 2147483647)
 
+    def test_decode_embedded_value_long_uint_bytes_consumed(self):
+        value = b'i\xff\xff\xff\xff'
+        self.assertEqual(decode._embedded_value(value)[0], 5)
+
+    def test_decode_embedded_value_long_uint_data_type(self):
+        value = b'i\xff\xff\xff\xff'
+        self.assertIsInstance(decode._embedded_value(value)[1], int)
+
+    def test_decode_embedded_value_long_uint_value(self):
+        value = b'i\xff\xff\xff\xff'
+        self.assertEqual(decode._embedded_value(value)[1], 4294967295)
+
     def test_decode_embedded_value_long_long_bytes_consumed(self):
         value = b'l\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertEqual(decode._embedded_value(value)[0], 9)
 
     @unittest.skipIf(PLATFORM_32BIT, 'Skipped on 32-bit platforms')
-    def test_decode_embedded_value_long_long_data_type(self):
+    def test_decode_embedded_value_long_long_data_type_64bit(self):
         value = b'l\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertIsInstance(decode._embedded_value(value)[1], int)
 
     @unittest.skipIf(PLATFORM_64BIT, 'Skipped on 64-bit platforms')
-    def test_decode_embedded_value_long_long_data_type(self):
+    def test_decode_embedded_value_long_long_data_type_32bit(self):
         value = b'l\x7f\xff\xff\xff\xff\xff\xff\xf8'
         self.assertIsInstance(decode._embedded_value(value)[1], long)
 
@@ -575,12 +641,27 @@ class CodecDecodeTests(unittest.TestCase):
         value = b's\x7f\xff'
         self.assertEqual(decode._embedded_value(value)[0], 3)
 
+    def test_decode_embedded_value_short_short_bytes_consumed(self):
+        self.assertEqual(decode._embedded_value(b'b\xff')[0], 2)
+
+    def test_decode_embedded_value_short_short_data_type(self):
+        self.assertIsInstance(decode._embedded_value(b'b\xff')[1], int)
+
+    def test_decode_embedded_value_short_short_value(self):
+        self.assertEqual(decode._embedded_value(b'b\xff')[1], 255)
+
     def test_decode_embedded_value_short_data_type(self):
         value = b's\x7f\xff'
         self.assertIsInstance(decode._embedded_value(value)[1], int)
 
     def test_decode_embedded_value_short_value(self):
         self.assertEqual(decode._embedded_value(b's\x7f\xff')[1], 32767)
+
+    def test_decode_embedded_value_short_uint_data_type(self):
+        self.assertIsInstance(decode._embedded_value(b'u\xff\xff')[1], int)
+
+    def test_decode_embedded_value_short_uint_value(self):
+        self.assertEqual(decode._embedded_value(b'u\xff\xff')[1], 65535)
 
     def test_decode_embedded_value_timestamp_bytes_consumed(self):
         value = b'T\x00\x00\x00\x00Ec)\x92'
@@ -627,9 +708,20 @@ class CodecDecodeTests(unittest.TestCase):
 
     def test_decode_embedded_value_field_table_value(self):
         value = decode._embedded_value(b'F' + self.FIELD_TBL)[1]
-        for key in value.keys():
+        for key in self.FIELD_TBL_VALUE.keys():
             if isinstance(value[key], float):
                 self.assertAlmostEqual(round(value[key], 3),
                                        round(self.FIELD_TBL_VALUE[key], 3))
             else:
                 self.assertEqual(value[key], self.FIELD_TBL_VALUE[key])
+
+    def test_decode_embedded_value_void_consumed(self):
+        self.assertEqual(decode._embedded_value(b'V')[0], 1)
+
+    def test_decode_embedded_value_void_value(self):
+        self.assertIsNone(decode._embedded_value(b'V')[1])
+
+    def test_field_embedded_value_field_table_keys(self):
+        value = decode._embedded_value(b'F' + self.FIELD_TBL)[1]
+        self.assertListEqual(sorted(value.keys()),
+                             sorted(self.FIELD_TBL_VALUE.keys()))